+++ OptimoveCore/Sources/Classes/Storage/FileStorage.swift
@@ -10,46 +10,41 @@ public protocol FileStorage {
     ///   - fileName: The file name on disk space.
     ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
     /// - Returns: Return `true` if the requested file exist.
-    func isExist(fileName: String, isGroupContainer: Bool) -> Bool
+    func isExist(fileName: String) -> Bool
 
     /// Save file.
     ///
     /// - Parameters:
     ///   - data: `Codable` object that will be saved.
     ///   - toFileName: The file name on disk space.
-    ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
-    func save<T: Codable>(data: T, toFileName: String, isGroupContainer: Bool) throws
+    func save<T: Codable>(data: T, toFileName: String) throws
 
     /// Save file.
     ///
     /// - Parameters:
     ///   - data: Raw data
     ///   - toFileName: The file name on disk space.
-    ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
-    func saveData(data: Data, toFileName: String, isGroupContainer: Bool) throws
+    func saveData(data: Data, toFileName: String) throws
 
     /// Load file.
     ///
     /// - Parameters:
     ///   - fileName: The file name on disk space.
-    ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
     /// - Returns: Return `Data` if file will be found, or `nil` if not.
-    func loadData(fileName: String, isGroupContainer: Bool) throws -> Data
+    func loadData(fileName: String) throws -> Data
 
     /// Load file.
     ///
     /// - Parameters:
     ///   - fileName: The file name on disk space.
-    ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
     /// - Returns: Return `Codable` if file will be found, or `nil` if not.
-    func load<T: Codable>(fileName: String, isGroupContainer: Bool) throws -> T
+    func load<T: Codable>(fileName: String) throws -> T
 
     /// Delete file.
     ///
     /// - Parameters:
     ///   - fileName: The file name on disk space.
-    ///   - isGroupContainer: set `true` if the requested file should be lookup in a group container.
-    func delete(fileName: String, isGroupContainer: Bool) throws
+    func delete(fileName: String) throws
 
 }
 
@@ -60,15 +55,11 @@ public final class FileStorageImpl {
     }
 
     private let fileManager: FileManager
-    private let groupedDirectoryURL: URL
-    private let sharedDirectoryURL: URL
+    private let directoryURL: URL
 
-    public init(bundleIdentifier: String,
-                fileManager: FileManager) throws {
-        self.fileManager = fileManager
-
-        groupedDirectoryURL = try fileManager.groupContainer(tenantBundleIdentifier: bundleIdentifier)
-        sharedDirectoryURL = try unwrap(fileManager.urls(for: .applicationSupportDirectory, in: .userDomainMask).first)
+    public init(url: URL) throws {
+        self.fileManager = FileManager.default
+        directoryURL = url
     }
 
     private func addSkipBackupAttributeToItemAtURL(fileURL: URL) throws {
@@ -78,22 +69,21 @@ public final class FileStorageImpl {
         try url.setResourceValues(resourceValues)
     }
 
-    private func getDirectory(isGroupContainer: Bool) -> URL {
-        let url = isGroupContainer ? groupedDirectoryURL : sharedDirectoryURL
-        return url.appendingPathComponent(Constants.folderName)
+    private func getDirectory() -> URL {
+        return directoryURL.appendingPathComponent(Constants.folderName)
     }
 }
 
 extension FileStorageImpl: FileStorage {
 
-    public func isExist(fileName: String, isGroupContainer: Bool) -> Bool {
-        let url = getDirectory(isGroupContainer: isGroupContainer)
+    public func isExist(fileName: String) -> Bool {
+        let url = getDirectory()
         let fileUrl = url.appendingPathComponent(fileName)
         return fileManager.fileExists(atPath: fileUrl.path)
     }
 
-    public func loadData(fileName: String, isGroupContainer: Bool) throws -> Data {
-        let fileUrl = getDirectory(isGroupContainer: isGroupContainer).appendingPathComponent(fileName)
+    public func loadData(fileName: String) throws -> Data {
+        let fileUrl = getDirectory().appendingPathComponent(fileName)
         do {
             let contents = try unwrap(fileManager.contents(atPath: fileUrl.path))
             Logger.debug("Load file at \(fileName)")
@@ -104,25 +94,23 @@ extension FileStorageImpl: FileStorage {
         }
     }
 
-    public func load<T: Codable>(fileName: String, isGroupContainer: Bool) throws -> T {
-        let data = try loadData(fileName: fileName, isGroupContainer: isGroupContainer)
+    public func load<T: Codable>(fileName: String) throws -> T {
+        let data = try loadData(fileName: fileName)
         return try JSONDecoder().decode(T.self, from: data)
     }
 
     public func save<T: Encodable>(
         data: T,
-        toFileName fileName: String,
-        isGroupContainer: Bool = false) throws {
+        toFileName fileName: String) throws {
         let data = try JSONEncoder().encode(data)
-        try saveData(data: data, toFileName: fileName, isGroupContainer: isGroupContainer)
+        try saveData(data: data, toFileName: fileName)
     }
 
     public func saveData(
         data: Data,
-        toFileName fileName: String,
-        isGroupContainer: Bool = false) throws {
+        toFileName fileName: String) throws {
         do {
-            let url = getDirectory(isGroupContainer: isGroupContainer)
+            let url = getDirectory()
             try fileManager.createDirectory(at: url, withIntermediateDirectories: true)
             let fileURL = url.appendingPathComponent(fileName)
             let success = fileManager.createFile(atPath: fileURL.path, contents: data, attributes: nil)
@@ -134,9 +122,9 @@ extension FileStorageImpl: FileStorage {
         }
     }
 
-    public func delete(fileName: String, isGroupContainer: Bool) throws {
+    public func delete(fileName: String) throws {
         do {
-            let fileUrl = getDirectory(isGroupContainer: isGroupContainer).appendingPathComponent(fileName)
+            let fileUrl = getDirectory().appendingPathComponent(fileName)
             try fileManager.removeItem(at: fileUrl)
             Logger.debug("File deleted at \(fileUrl.absoluteString).")
         } catch {