+++ OptimoveSDK/Sources/Classes/Objects/User.swift
@@ -1,0 +1,14 @@
+//  Copyright © 2020 Optimove. All rights reserved.
+
+import Foundation
+
+final class User {
+
+    let userID: String
+    let visitorID: String
+
+    init(userID: String) {
+        self.userID = userID
+        self.visitorID = userID.sha1().prefix(16).description.lowercased()
+    }
+}