+++ OptimoveSDK/Sources/Classes/OnStartHelpers/OnStartEventGenerator.swift
@@ -19,14 +19,13 @@ final class OnStartEventGenerator {
 
     func generate() {
         asyncGenerate()
-        generateEvents()
+        syncGenerate()
     }
 
-    private func generateEvents() {
+    private func syncGenerate() {
         tryCatch {
             let metaDataEvent = try coreEventFactory.createEvent(.metaData)
-            let setAdvertisingIdEvent = try coreEventFactory.createEvent(.setAdvertisingId)
-            self.synchronizer.handle(.report(events: [metaDataEvent, setAdvertisingIdEvent]))
+            self.synchronizer.handle(.report(events: [metaDataEvent]))
         }
     }